refactor(slack-bot): extract intent.js from index.js (behavior-preserving)
